Addition to the removal of trailing whitespaces
By putting the substitution that you had before into this function, when the
function is called it will also save your current cursor position before the
substitution is made. This way, after a substitution is made on a save your
cursor will not be at the bottom of the file where the last substitution was
made.

I also have:

```vim
au BufWritePre * :call...
```

instead of:

```vim
au BufWritePre *.rb :call...
```

Because I like to remove trailing whitespace from all files I save with vim, not
just ruby files. But if you prefer to only have that take effect in ruby files
then that line can be changed to reflect that.